move unused methods out of Document
